requirements.txt
investigate==1.2.4
shodan==1.3.3
geoip2==2.2.0
lxml==3.4.4
pymongo==3.1.1
colorama==0.3.3
dnspython==1.12.0
pyzmq==16.0.2
whois==0.7
selenium==3.0.2
requests==2.11.1
maxminddb==1.2.1
falcon==1.0.0
redis==2.10.5
